Add files via upload
Data Science is a competitive field which has various Job roles. In the quest to know the salary ranges of all fields in data science, I came up with the idea to prepare a dashboard that shows each field and their salary ranges with respect to their job-type. The dataset used in this project was gotten from Kaggle.